Final merge

See merge request cse2115/2023-2024/group-11/team-11a!47

+++ order-microservice/src/main/java/nl/tudelft/sem/template/order/controllers/OrderController.java
@@ -8,11 +8,21 @@ import nl.tudelft.sem.template.model.Address;
 import nl.tudelft.sem.template.model.Order;
 import nl.tudelft.sem.template.order.domain.helpers.FilteringByStatus;
 import nl.tudelft.sem.template.order.domain.helpers.FilteringParam;
+import nl.tudelft.sem.template.order.domain.helpers.OrderValidation;
+import nl.tudelft.sem.template.order.domain.user.CustomerNotFoundException;
+import nl.tudelft.sem.template.order.domain.user.DishNotFoundException;
+import nl.tudelft.sem.template.order.domain.user.InvalidOrderRatingException;
+import nl.tudelft.sem.template.order.domain.user.InvalidOrderStatusException;
 import nl.tudelft.sem.template.order.domain.user.NoOrdersException;
 import nl.tudelft.sem.template.order.domain.user.NullFieldException;
 import nl.tudelft.sem.template.order.domain.user.OrderNotFoundException;
 import nl.tudelft.sem.template.order.domain.user.OrderService;
+import nl.tudelft.sem.template.order.domain.user.UserIDNotFoundException;
+import nl.tudelft.sem.template.order.domain.user.VendorNotFoundException;
+import nl.tudelft.sem.template.user.services.JsonParserService;
+import nl.tudelft.sem.template.user.services.UserMicroServiceService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -20,10 +30,22 @@ import org.springframework.web.bind.annotation.RestController;
 public class OrderController implements OrderApi {
 
     private final transient OrderService orderService;
+    private final transient UserMicroServiceService userMicroServiceService;
+    private final transient DishController dishController;
 
+    /**
+     * Constructor method for OrderController.
+     *
+     * @param orderService an orderservice
+     * @param userMicroServiceService a userMicroServiceService
+     * @param dishController a dishController
+     */
     @Autowired
-    public OrderController(OrderService orderService) {
+    public OrderController(OrderService orderService, UserMicroServiceService userMicroServiceService,
+                           DishController dishController) {
         this.orderService = orderService;
+        this.userMicroServiceService = userMicroServiceService;
+        this.dishController = dishController;
     }
 
     /**
@@ -49,15 +71,25 @@ public class OrderController implements OrderApi {
 
     /**
      * Endpoint for returning all Orders in the database.
+     * First checks if the user is an admin
      *
      * @return 200 OK - The Orders are successfully returned
+     *         401 UNAUTHORIZED - The user is not authorised to ask for all the orders
      *         404 NOT FOUND - No Orders are stored in the database
      */
     @Override
-    public ResponseEntity<List<Order>> getAllOrders() {
+    public ResponseEntity<List<Order>> getAllOrders(UUID userID) {
+
         try {
-            List<Order> list = orderService.getAllOrders();
-            return ResponseEntity.ok(list);
+            String jsonUser = userMicroServiceService.getUserInformation(userID);
+            userNotFoundThrower(userID, jsonUser);
+            String userType = JsonParserService.parseUserType(jsonUser);
+            if (userType.equals("Admin")) {
+                List<Order> list = orderService.getAllOrders();
+                return ResponseEntity.ok(list);
+            } else {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+            }
         } catch (Exception e) {
             return ResponseEntity.notFound().build();
         }
@@ -68,6 +100,7 @@ public class OrderController implements OrderApi {
      *
      * @param orderID The provided ID of the order to return
      * @return 200 OK - The provided Order is returned
+     *         422 UNPROCESSABLE ENTITY if the orderID is null
      *         400 BAD REQUEST - Returning the provided Order is unsuccessful
      *         404 NOT FOUND - No Order exists with the provided ID
      */
@@ -84,7 +117,26 @@ public class OrderController implements OrderApi {
         } catch (Exception e) {
             return ResponseEntity.badRequest().build();
         }
+    }
 
+    /**
+     * Endpoint for returning an order to a vendor by specifying its id.
+     * The order is returned only after it is paid and if all of its dishes are available
+     *
+     * @param orderId the id of the order to be retrieved
+     * @return 200 OK if the order has been successfully retrieved, including the order
+     *         404 NOT FOUND if the order could not be found
+     *         400 BAD REQUEST if the retrieval was not successful or the order was invalid
+     */
+    @Override
+    public ResponseEntity<Order> orderOrderIDVendorGet(UUID orderId) {
+        OrderValidation orderValidation = new OrderValidation(this, dishController);
+        ResponseEntity<Boolean> response = orderValidation.isOrderValid(orderId);
+        if (Boolean.TRUE.equals(response.getBody())) {
+            ResponseEntity<Order> order = getOrderById(orderId);
+            return order;
+        }
+        return ResponseEntity.notFound().build();
     }
 
     /**
@@ -97,23 +149,55 @@ public class OrderController implements OrderApi {
      *         404 NOT FOUND - No Order exists with the provided ID
      */
     @Override
-    public ResponseEntity<Order> editOrderByID(UUID orderID, Order order) {
+    public ResponseEntity<Order> editOrderByID(UUID orderID, UUID userID, Order order) {
 
         if (!order.getOrderID().equals(orderID)) {
             return ResponseEntity.badRequest().build();
         }
 
         try {
-            Order edited = orderService.editOrderByID(orderID, order);
-            return ResponseEntity.ok(edited);
-        } catch (NullFieldException e) {
-            return ResponseEntity.unprocessableEntity().build();
+            String jsonUser = userMicroServiceService.getUserInformation(userID);
+            userNotFoundThrower(userID, jsonUser);
+            String userType = JsonParserService.parseUserType(jsonUser);
+            return adminCheckerEdit(orderID, order, userType);
         } catch (OrderNotFoundException e) {
             return ResponseEntity.notFound().build();
         } catch (Exception e) {
             return ResponseEntity.badRequest().build();
         }
+    }
 
+    /** Throws exception if the id is malformed.
+     *
+     * @param userID the id to show in exception
+     * @param jsonUser the json to check on
+     * @throws UserIDNotFoundException the exception to throw
+     */
+    private static void userNotFoundThrower(UUID userID, String jsonUser) throws UserIDNotFoundException {
+        if (jsonUser == null || jsonUser.isEmpty()) {
+            throw new UserIDNotFoundException(userID);
+        }
+    }
+
+    /** Checks if the usertype is an admin.
+     *
+     * @param orderID the id of the order
+     * @param order the whole object
+     * @param userType the current type
+     * @return different status code based on input
+     * @throws OrderNotFoundException when the order is not found
+     * @throws NullFieldException if the order is null
+     * @throws VendorNotFoundException if the vendor id is wrong
+     * @throws CustomerNotFoundException if the customer id is wrong
+     */
+    private ResponseEntity<Order> adminCheckerEdit(UUID orderID, Order order, String userType)
+            throws OrderNotFoundException, NullFieldException, VendorNotFoundException, CustomerNotFoundException {
+        if (userType.equals("Admin")) {
+            Order edited = orderService.editOrderByID(orderID, order);
+            return ResponseEntity.ok(edited);
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
     }
 
     /**
@@ -125,14 +209,32 @@ public class OrderController implements OrderApi {
      *         404 NOT FOUND - No Order exists with the provided ID
      */
     @Override
-    public ResponseEntity<Void> deleteOrderByID(UUID orderID) {
+    public ResponseEntity<Void> deleteOrderByID(UUID orderID, UUID userID) {
         try {
-            orderService.deleteOrderByID(orderID);
-            return ResponseEntity.ok().build();
+            String jsonUser = userMicroServiceService.getUserInformation(userID);
+            userNotFoundThrower(userID, jsonUser);
+            String userType = JsonParserService.parseUserType(jsonUser);
+            return adminChecker(orderID, userType);
         } catch (OrderNotFoundException e) {
             return ResponseEntity.notFound().build();
         } catch (Exception e) {
             return ResponseEntity.badRequest().build();
+        }
+    }
+
+    /** Admin checker helper function.
+     *
+     * @param orderID the id of the current order
+     * @param userType admin or not admin
+     * @return the status code of this operation
+     * @throws OrderNotFoundException if the order does not exist
+     */
+    private ResponseEntity<Void> adminChecker(UUID orderID, String userType) throws OrderNotFoundException {
+        if (userType.equals("Admin")) {
+            orderService.deleteOrderByID(orderID);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
     }
 
@@ -158,6 +260,7 @@ public class OrderController implements OrderApi {
             return ResponseEntity.badRequest().build();
         }
     }
+
     /**
      * OrderID isPaid controller method to update the isPaid field.
      * It throws a 404 if the order is not found.
@@ -165,7 +268,6 @@ public class OrderController implements OrderApi {
      * @param orderID the id of the order to be checked
      * @return the order after it was updated
      */
-
     @Override
     public ResponseEntity<Order> updateOrderPaid(UUID orderID) {
         try {
@@ -184,6 +286,7 @@ public class OrderController implements OrderApi {
      *
      * @param orderID the id of the order to be retrieved the list of dishes from.
      * @return 200 OK if the list of dishes has been retrieved, including the list of UUID dishes
+     *         422 UNPROCESSABLE ENTITY if the orderID is null
      *         404 NOT FOUND if the order could not be found
      *         400 BAD REQUEST if the retrieval was not successful
      */
@@ -207,6 +310,7 @@ public class OrderController implements OrderApi {
      *
      * @param orderID the id of the order to be retrieved the special requirements from.
      * @return 200 OK if the special requirements of the order has been retrieved, including the special requirements
+     *         422 UNPROCESSABLE ENTITY if the orderID is null
      *         404 NOT FOUND if the order could not be found
      *         400 BAD REQUEST if the retrieval was not successful
      */
@@ -229,6 +333,7 @@ public class OrderController implements OrderApi {
      *
      * @param orderID the id of the order to be delivered to the retrieved address.
      * @return 200 OK if the address of the order has been retrieved, including the address
+     *         422 UNPROCESSABLE ENTITY if the orderID is null
      *         404 NOT FOUND if the order could not be found
      *         400 BAD REQUEST if the retrieval was not successful
      */
@@ -251,6 +356,7 @@ public class OrderController implements OrderApi {
      *
      * @param orderID the id of the order of which the date of creation should be retrieved.
      * @return 200 OK if the date of the order has been retrieved, including the date
+     *         422 UNPROCESSABLE ENTITY if the orderID is null
      *         404 NOT FOUND if the order could not be found
      *         400 BAD REQUEST if the retrieval was not successful
      */
@@ -260,6 +366,32 @@ public class OrderController implements OrderApi {
             return ResponseEntity.ok(order.getDate());
         } catch (NullFieldException e) {
             return ResponseEntity.unprocessableEntity().build();
+        } catch (OrderNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
+        }
+    }
+
+    /**
+     * OrderID getCustomerName controller method.
+     * It returns the name of the customer that created a specific order
+     *
+     * @param orderID the id of the order from which the customer name should be retrieved
+     * @return 200 OK if the name of the customer has been successfully retrieved, including the name
+     *         422 UNPROCESSABLE ENTITY if the orderID is null
+     *         404 NOT FOUND if the order or the user could not be found
+     *         400 BAD REQUEST if the retrieval was not successful
+     */
+    public ResponseEntity<String> getCustomerName(UUID orderID) {
+        try {
+            Order order = orderService.getOrderById(orderID);
+            UUID userID = order.getCustomerID();
+            return ResponseEntity.ok(userMicroServiceService.getUserName(userID));
+        } catch (NullFieldException e) {
+            return ResponseEntity.unprocessableEntity().build();
+        } catch (UserIDNotFoundException userIDNotFoundException) {
+            return ResponseEntity.notFound().build();
         } catch (OrderNotFoundException e) {
             return ResponseEntity.notFound().build();
         } catch (Exception e) {
@@ -283,6 +415,159 @@ public class OrderController implements OrderApi {
             return ResponseEntity.ok(allOrdersByCustomerID);
         } catch (NoOrdersException noOrdersException) {
             return ResponseEntity.notFound().build();
+        } catch (CustomerNotFoundException customerNotFoundException) {
+            return ResponseEntity.notFound().build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
+        }
+    }
+
+    /**Controller for the /order/{orderID}/totalCost endpoint.
+     * It takes the implementation provided in VendorAnalyticsController.
+     *
+     * @param orderID ID of order that needs to be fetched (required)
+     * @return a response where for 200 returns the total sum of the
+     *          dishes that can be found in the database
+     */
+    @Override
+    public ResponseEntity<Float> orderOrderIDTotalCostGet(UUID orderID) {
+        VendorAnalyticsController vendorAnalyticsController =
+                new VendorAnalyticsController(this, dishController, orderService);
+        return vendorAnalyticsController.getOrderEarnings(orderID);
+    }
+
+    /**
+     * Controller for the /order/{orderID}/addDishToOrder/{dishID}.
+     *
+     * @param orderID the id of the order to add the dish to
+     * @param dishID  the id of the dish to be added to the order
+     * @return 200 OK if the dish has been added successfully to the order, including the order
+     *         422 UNPROCESSABLE ENTITY if the orderID or dishID are null
+     *         404 NOT FOUND if the order or the dish could not be found
+     *         400 BAD REQUEST if the operation was not successful
+     */
+    @Override
+    public ResponseEntity<Order> addDishToOrder(UUID orderID, UUID dishID) {
+        try {
+            Order order = orderService.addDishToOrder(orderID, dishID);
+            return ResponseEntity.ok(order);
+        } catch (NullFieldException e) {
+            return ResponseEntity.unprocessableEntity().build();
+        } catch (OrderNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        } catch (DishNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
+        }
+    }
+
+    /**
+     * Controller for the /order/{orderID}/removeDishFromOrder/{dishID}.
+     *
+     * @param orderID the id of the order to remove the dish from
+     * @param dishID  the id of the dish to be removed from the order
+     * @return 200 OK if the dish has been removed successfully from the order, including the order
+     *         422 UNPROCESSABLE ENTITY if the orderID or dishID are null
+     *         404 NOT FOUND if the order or the dish could not be found
+     *         400 BAD REQUEST if the operation was not successful
+     */
+    @Override
+    public ResponseEntity<Order> removeDishFromOrder(UUID orderID, UUID dishID) {
+        try {
+            Order order = orderService.removeDishFromOrder(orderID, dishID);
+            return ResponseEntity.ok(order);
+        } catch (NullFieldException e) {
+            return ResponseEntity.unprocessableEntity().build();
+        } catch (OrderNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        } catch (DishNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
+        }
+
+    }
+
+    /**
+     * Endpoint for getting the status of an Order.
+     *
+     * @param orderID ID of order to return (required)
+     * @return status of Order in String value
+     */
+    @Override
+    public ResponseEntity<String> getStatusOfOrderById(UUID orderID) {
+
+        try {
+            String status = orderService.getStatusOfOrderById(orderID).toString();
+            return ResponseEntity.ok(status);
+        } catch (OrderNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
+        }
+    }
+
+    /**
+     * Endpoint for updating the status of an Order.
+     *
+     * @param orderID ID specifying the Order
+     * @param status New Status as String
+     * @return Edited Order in the database
+     */
+    @Override
+    public ResponseEntity<Void> updateStatusOfOrderById(UUID orderID, String status) {
+        try {
+            orderService.updateStatusOfOrderById(orderID, status);
+            return ResponseEntity.ok().build();
+        } catch (OrderNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        } catch (InvalidOrderStatusException e) {
+            return ResponseEntity.status(HttpStatus.UNSUPPORTED_MEDIA_TYPE).build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
+        }
+
+    }
+
+    /**
+     * Endpoint for getting the rating of an Order.
+     *
+     * @param orderID ID specifying order
+     * @return rating of Order as int value
+     */
+    @Override
+    public ResponseEntity<Integer> getOrderRatingByID(UUID orderID) {
+
+        try {
+            int rating = orderService.getOrderRatingByID(orderID);
+            return ResponseEntity.ok(rating);
+        } catch (OrderNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
+        }
+    }
+
+
+    /**
+     * Endpoint for updating the rating of an Order.
+     *
+     * @param orderID ID specifying the Order
+     * @param rating New rating as an Integer
+     * @return ResponseEntity of operation
+     */
+    @Override
+    public ResponseEntity<Order> editOrderRatingByID(UUID orderID, Integer rating) {
+        try {
+            Order o = orderService.editOrderRatingByID(orderID, rating);
+            return ResponseEntity.ok(o);
+        } catch (OrderNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        } catch (InvalidOrderRatingException e) {
+            return ResponseEntity.badRequest().build();
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().build();
         }
     }
 }
